+++ lib/src/models/unwrapped_gift_wrap.dart
@@ -1,0 +1,7 @@
+import 'package:ndk/ndk.dart';
+
+class UnwrappedGiftWrap {
+  final Nip01Event seal;
+  final Nip01Event rumor;
+  UnwrappedGiftWrap({required this.seal, required this.rumor});
+}